django_project/requirements.txt
Django==4.2.2
django-environ==0.10.0
psycopg2-binary==2.9.6
djangorestframework==3.14.0
django-filter==23.2
drf-yasg==1.21.6
djoser==2.2.0
coreapi==2.3.3
